moved most of the RESP logic into base, resp classes
